+++ energy_view.h
@@ -126,7 +126,7 @@ private:
 	void setBannerPage(int status, const QDate &selection_date);
 	QString dateToKey(const QDate &date, EnergyDevice::GraphType t);
 	void updateBanners();
-	void updateBanner(Bann2Buttons *banner, unsigned int value, int dec, QString symbol);
+	void updateBanner(Bann2Buttons *banner, qint64 value, int dec, QString symbol);
 	void updateCurrentGraph();
 	// returns true if we own the graph/table instances
 	bool isGraphOurs();
@@ -140,8 +140,8 @@ private:
 	bannFrecce *bannNavigazione;
 	Bann2Buttons *current_banner, *daily_av_banner;
 	Bann2Buttons *cumulative_day_banner, *cumulative_month_banner, *cumulative_year_banner;
-	unsigned int current_value, daily_av_value;
-	unsigned int cumulative_day_value, cumulative_month_value, cumulative_year_value;
+	qint64 current_value, daily_av_value;
+	qint64 cumulative_day_value, cumulative_month_value, cumulative_year_value;
 	TimePeriodSelection *time_period;
 	QStackedWidget *widget_container;
 	EnergyTable *table;
